feat: event 핸들러 추가
